[svn-r27993] Merge of r27805 from trunk.
Tested on: Ubuntu 15.04 (Linux 3.19 x86_64) gcc 4.9.2

+++ src/H5Cpkg.h
@@ -821,7 +821,7 @@ if ( ( (entry_ptr) == NULL ) ||                                                \
 
 #define H5C__HASH_MASK		((size_t)(H5C__HASH_TABLE_LEN - 1) << 3)
 
-#define H5C__HASH_FCN(x)	(int)(((x) & H5C__HASH_MASK) >> 3)
+#define H5C__HASH_FCN(x)	(int)((unsigned)((x) & H5C__HASH_MASK) >> 3)
 
 #if H5C_DO_SANITY_CHECKS
